`storage`: replace book-keeping types with `ssize_t`
The assumption that the result from a compaction operation will always
result in a positive number (i.e `bytes_after < bytes_before`) was an
invalid one.

For sake of ease and correctness, replace the book-kept variables
(previously `uint64_t`) with their signed equivalents (`ssize_t`).

Also, add a `dassert()` to `subtract_closed/dirty_bytes()` to ensure
their values do not become negative.

+++ src/v/raft/tests/failure_injectable_log.cc
@@ -240,11 +240,11 @@ failure_injectable_log::retention_offset(storage::gc_config cfg) const {
     return _underlying_log->retention_offset(cfg);
 }
 
-uint64_t failure_injectable_log::dirty_segment_bytes() const {
+ssize_t failure_injectable_log::dirty_segment_bytes() const {
     return _underlying_log->dirty_segment_bytes();
 }
 
-uint64_t failure_injectable_log::closed_segment_bytes() const {
+ssize_t failure_injectable_log::closed_segment_bytes() const {
     return _underlying_log->closed_segment_bytes();
 }
